Update name_mother.html
When I submit name_mother.html I get the code for name_mother.php instead of the bedtime story I want.

Also not sure how to have the php output in on an html page. Stick the php inside body tags? Return a value for the story and use that in an html section?

Can't test until I get the submit to work properly